Update mapping handling; add gc
Eventually we could optionally use custom fields instead of
storing our map in a hash on our config card.

The current approach has a problem where if you archive a card,
then gc, then unarchive it, it will be assigned a new Trolly
ID.  However, because I assume the amount of data which can be
stored in the description of a trello card is finite, we can't
just have it being unbounded.

Signed-off-by: Lon Hohberger <lhh@redhat.com>