Rename project to jirate
Instances of the "trolly" name intentionally left alone:
- CLI entry point.
- Trello-specific environment variables.
- The TrollyBoard class and related bits.
- The github URL, which we can change when we're ready to rename
  the actual github project.

Trello modules continue to be first-class, with Jira ones still
relegated to jira_* filenames. We could change that in the future,
if there ends up being a lot more development on that side or we decide
to split Trello code out to its own project.