jira: Add a custom field mapping resolver
The way the jira module for python is implemented,
there's no simple way to feed the server's custom
field definitions to produce something human-readable
that can be used in code, i.e.:

  issue.fields.fixed_in_build
    vs.
  issue.fields.custom_321789045789

or:

  issue.raw['fields']['fixed_in_build']
    vs.
  issue.raw['fields']['custom_321789045789']

The raw JSON returned from the JIRA server for an issue
does not contain any of the mapping information, which
must be retrieved separately. Fortunately, this changes
infrequently, so we can cache it in the Jirate() object
for now, although if we want this to work for long-running
processes, it should be updated periodically.

So, we're left with:

   jirate.field(issue, 'fixed_in_build')
     or
   issue.field('fixed_in_build')

... which lazy-loads the field definitions and uses
a combination of toolchest's nym() function and the raw
field names to try to find the field the caller was
looking for.

Additional notes:

* Never do a round-trip to the JIRA server when called from
the Issue object.

* Do not overwrite jira.resources.Issue.field if it exists.

* Raise AttributeError if a field does not exist.