jira: Don't index all open issues by default
In the original Trello code, the default was to index all open cards.
This model does not work for large JIRA projects and is uniquely
unkind to the JIRA server in that case.

A future patch will allow this to be a configuration option, but,
for now, change the default to only search for issues assigned to
a user and index those.