Add field translation functions
The existing field_map() function was able to translate human-readable
names into field IDs. Renamed that to field_to_id() and added similar
functions to similarly translate to full names and to aliases as well.

This will make it easier to work with fields when we read user input or
write output that humans as supposed to parse easily.