Field rendering: allow returning objects instead of strings
This allows us to use a lot of the field parsing code to process issue
data before printing it, such as for template generation.

Existing behavior is preserved for the regular strings-for-printing
case.

Note some fields don't really have special object handling yet, such as
dates and usernames. We can add that if/when we actually need it.