decor: truncate field name to max width of field values to save horizontal space
Some field names have considerably longer names than the values they contain.
For example, "Story Points" is usually a numerical value with the longest value
being unset ("None").  Rendering a matrix like the following (fields:
"story_points,summary"):

  key    | story_points | summary
  TEST-1 | 1            | test
  TEST-2 | 11           | test2

This is a waste of horizontal screen real estate.  Instead, truncate the field
name width to the maximum value of the column to save space:

  key    | s… | summary
  TEST-1 | 1  | test
  TEST-2 | 11 | test2

This also changes a single-character width from '…' to the first character of
the field name:

  key    | s | summary
  TEST-1 | 1 | test