(dirty) Allow sprint setting by ID
The Jira REST APIs for sprints are an entirely different API
tree (/rest/sprint/1.0 vs /rest/api/latest/...).

Until a major feature is added to support the new API tree
and using it properly, this can be used as a quick and dirty
way to add an issue to a sprint, i.e.:

   $ jirate field MYKEY-123 set sprint 12345

Unfortunately, there's no way to set via "name".