Fix block quote handling
When generating a template from an issue, if a line in the
description of the issue or a subtask had a trailing space,
python yaml would encode the whole thing as a quoted string
instead of a block-quote like it should have been.

By iterating through the text and stripping off trailing
whitespace, first on the whole, then on each line individually,
this issue is resolved.

Related to: https://github.com/yaml/pyyaml/issues/121

Found / original patch by Jon Schlueter