jira: Ignore values that start with underscore in templates
If a user wants to set a variable in a jinja2 template that
is the parsed value of something else, there was no way to
do that.

This allows jirate templates to set other variables without
jirate requiring a value be passed ahead of time (they
must start with underscore).

* It's a user error to set a variable that starts with an underscore
* Underscore variables do not require values from the user

Closes #77