CLI BREAK: automatically print the active sprint, if there is exactly one
This could be controversial, but having to remember a sprint ID
in all cases is not user-focused.  This changes the behavior
of sprint_info() to list the issues in the sole active sprint,
if querying the sprint table nets one active sprint. Else, it prints
the sprint list like before.

A new flag (-l, --list) is added to always list current and future
sprints.